requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
google-adk>=1.0.0
google-genai>=1.0.0
google-cloud-firestore>=2.19.0
google-auth>=2.36.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.150.0
pydantic>=2.10.0
httpx>=0.28.0
